Updated Shop and BuildManager
Towers now Cost money (int). Towers can be placed if player has enough money.

+++ Tower Defense Game/Assets/Scripts/IPurchasble.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IPurchasble
 {
-    float GetTowerCost();
+    int GetTowerCost();
 }
